% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cosine_similarity.R
\name{compute_cosine_similarity}
\alias{compute_cosine_similarity}
\title{Compute Cosine Similarity}
\usage{
compute_cosine_similarity(
  dataframe,
  doc_col,
  content_col,
  output_type = "matrix"
)
}
\arguments{
\item{dataframe}{A dataframe containing the documents and content.}

\item{doc_col}{The name of the column in the dataframe that contains document identifiers.}

\item{content_col}{The name of the column that contains the document content.}

\item{output_type}{The type of output required: "matrix" or "edgelist".}
}
\value{
A matrix or edge list representing the cosine similarities between documents.
}
\description{
This function calculates the cosine similarity between documents in a dataframe.
The output can be returned as either a matrix or an edge list.
}
\examples{
\dontrun{
# Example dataframe with 'id' and 'text' columns
df <- data.frame(id = c(1, 2, 3), text = c("text one", "text two", "text three"))
# Compute cosine similarity and return a matrix
similarity_matrix <- compute_cosine_similarity(df, "id", "text", "matrix")
}
}
