% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist_from_entities.R
\name{edgelist_from_entities}
\alias{edgelist_from_entities}
\title{Generate Edge List or Matrix from Entity Pairs in Text}
\usage{
edgelist_from_entities(
  df,
  text_var,
  id_var,
  entities_list,
  output_type = "edgelist"
)
}
\arguments{
\item{df}{A dataframe with text documents and their identifiers.}

\item{text_var}{The column in df containing text content.}

\item{id_var}{The column in df containing document identifiers.}

\item{entities_list}{A named list where each element is a vector of synonyms for an entity.}

\item{output_type}{The desired output: "edgelist" (default) or "matrix".}
}
\value{
If output_type is "edgelist", a dataframe with columns for document_id, source entity, and target entity.
If output_type is "matrix", a matrix of co-occurrence counts.
}
\description{
Processes a dataframe containing text content and identifiers, as well as a list of entities,
to create an edge list or matrix based on the co-occurrence of entities within the same document.
}
\examples{
\dontrun{
text_df <- data.frame(
  id = 1:3,
  text_content = c(
    "Acset Indonusa and Adhi Commuter Property",
    "ABM and ACE",
    "Astra Agro Lestari and Ace Hardware Indonesia collaborate"
  ),
  stringsAsFactors = FALSE
)

entities_list <- list(
  AALI = c("Astra Agro Lestari", "Astra Agro"),
  ABMM = c("ABM Investama", "ABM"),
  ACES = c("Ace Hardware Indonesia", "Ace Hardware", "Ace"),
  ACST = c("Acset Indonusa", "Acset"),
  ADCP = c("Adhi Commuter Properti", "Adhi")
)

edges <- edgelist_from_entities(text_df, "text_content", "id", entities_list, "edgelist")
matrix <- edgelist_from_entities(text_df, "text_content", "id", entities_list, "matrix")
print(edges)
print(matrix)
}
}
