% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist_from_group.R
\name{edgelist_from_group}
\alias{edgelist_from_group}
\title{Create an edge list from a dataframe based on groups}
\usage{
edgelist_from_group(df, group_col, item_col, remove_duplicates = FALSE)
}
\arguments{
\item{df}{A dataframe with at least two columns.}

\item{group_col}{The name of the column that contains the grouping variable.}

\item{item_col}{The name of the column that contains the items to be connected within groups.}

\item{remove_duplicates}{Logical; if TRUE, removes duplicate edges.}
}
\value{
A dataframe with two columns representing the source and target of each edge.
}
\description{
This function takes a dataframe and constructs an edge list where each item within the same group is connected.
It can optionally remove duplicate edges.
}
\examples{
df <- data.frame(
  Group = c('A', 'A', 'B', 'B', 'B', 'C'),
  Item = c('Item1', 'Item2', 'Item1', 'Item2', 'Item3', 'Item1')
)
edge_list <- edgelist_from_group(df, 'Group', 'Item', remove_duplicates = TRUE)
print(edge_list)

}
