% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist_from_text.R
\name{edgelist_from_text}
\alias{edgelist_from_text}
\title{Create an edge list from a text using a sliding window}
\usage{
edgelist_from_text(
  text = NULL,
  window_size = 1,
  remove_duplicates = FALSE,
  stopwords = character(0)
)
}
\arguments{
\item{text}{A character string from which to create the edge list.}

\item{window_size}{An integer defining the size of the sliding window.}

\item{remove_duplicates}{Logical; if TRUE, removes duplicate edges.}

\item{stopwords}{A character vector of words to be removed from the text.}
}
\value{
A data frame with two columns representing the source and target of each edge.
}
\description{
This function takes a string of text and creates an edge list based on a sliding window.
It can optionally remove duplicate edges and exclude specified stopwords.
}
\examples{
text <- "I like you in hate"
stopwords <- c("in")
edge_list <- edgelist_from_text(text, 3, remove_duplicates = TRUE, stopwords = stopwords)
print(edge_list)

}
