% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist_from_text.R
\name{edgelist_from_texts}
\alias{edgelist_from_texts}
\title{Generate Edge List from Text with Context Window}
\usage{
edgelist_from_texts(
  texts = NULL,
  window_size = 2,
  remove_duplicates = FALSE,
  stopwords = character(0)
)
}
\arguments{
\item{window_size}{A positive integer indicating the size of the context window around each word for edge formation (default is 1).}

\item{remove_duplicates}{Logical; if TRUE, duplicate edges are removed, considering that an edge between two words is undirected.}

\item{stopwords}{A character vector of stopwords to be removed from the text (default is an empty character vector).}

\item{text}{A string containing the text to be processed.}
}
\value{
A dataframe with two columns 'source' and 'target', each row representing an edge between two words in the text within the context window.
}
\description{
Processes a string of text and generates an edge list where each edge represents a pair of words that appear within a certain distance defined by the window size. The function allows for the removal of stopwords and duplicates to reduce noise in the edge list.
}
\examples{
\dontrun{
sample_text <- "I like programming in R"
sample_stopwords <- c("in")
# Generate edge list with a window size of 4 and remove duplicates
edge_list <- edgelist_from_text(sample_text, 4, TRUE, sample_stopwords)
print(edge_list)
}
}
